Move AutoImportQueue into Meteor directory.
