+++ npm-packages/meteor-vite/src/vite/ViteLoadRequest.ts
@@ -6,7 +6,7 @@ import pc from 'picocolors';
 import { ViteDevServer } from 'vite';
 import { createLabelledLogger, LabelLogger } from '../utilities/Logger';
 import { isSameModulePath } from '../meteor/package/components/MeteorPackage';
-import AutoImportQueue from './AutoImportQueue';
+import AutoImportQueue from '../meteor/package/AutoImportQueue';
 import { MeteorViteError } from '../error/MeteorViteError';
 import type { PluginSettings } from './plugin/MeteorStubs';
