+++ npm-packages/meteor-vite/test/AutoImportQueue.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, test } from 'vitest';
-import AutoImportQueue, { wait } from '../src/vite/AutoImportQueue';
+import AutoImportQueue, { wait } from '../src/meteor/package/AutoImportQueue';
 import { RefreshNeeded } from '../src/vite/ViteLoadRequest';
 import { AutoImportMock } from './__mocks';
